+++ HydroSolver/Reservoir/NumericalParameters.cpp
@@ -11,7 +11,7 @@ namespace reservoir_simulator
 		: newtonTol(nTol), newtonMaxIterNmbr(nCount), AMG_AbsTol(amg_AbsTol), AMG_RelTol(amg_RelTol) {}
 	NumericalParameters::NumericalParameters() noexcept { }
 
-	void NumericalParameters::update_currentAMGState(const std::tuple<int, double, bool>& AMGstate)
+	void NumericalParameters::update_currentAMGState(const std::tuple<size_t, double, bool>& AMGstate)
 	{
 		AMG_currentIterationCount = std::get<0>(AMGstate);
 		double newAMG_error = std::get<1>(AMGstate);
